requirements.txt
Brotli
certifi
cffi
charset-normalizer
colorama
coloredlogs==15.0.1
contourpy
cycler
filelock
flatbuffers==23.3.3
fonttools
gmpy2 
h2 
hpack==4.0.0
humanfriendly==10.0
hyperframe 
idna 
importlib_resources 
Jinja2 
kiwisolver 
lsq-ellipse==2.2.1
MarkupSafe 
matplotlib  
mpmath 
munkres==1.1.4
networkx 
numpy 
opencv-python==4.10.0.84
packaging 
Pillow 
ply 
pycparser 
pyparsing 
PyQt5==5.15.9
PyQt5-sip==12.12.2
PySocks 
python-dateutil 
PyYAML 
requests 
scipy==1.13.1
shapely 
sip 
six 
sympy 
toml 
tomli 
tornado 
tqdm 
triton==2.3.1
typing_extensions 
unicodedata2 
urllib3
zipp
zstandard==0.22.0
ultralytics
#samurai installation: !git clone https://github.com/yangchris11/samurai.git
#PNLcalib installation